requirements.txt
anthropic==0.7.7
python-dotenv==1.0.0
pyyaml==6.0.1